requirements.txt
# Core PDF parsing
pdfplumber==0.10.3
pypdf==4.2.0

# Vector store
chromadb==0.5.0

# Embeddings (local, no API needed)
sentence-transformers==3.0.1

# LangChain agent framework
langchain==0.2.6
langchain-anthropic==0.1.19
langchain-community==0.2.6
langchain-chroma==0.1.2

# API server
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4

# Utilities
python-dotenv==1.0.1
pandas==2.2.2
tabulate==0.9.0
numpy==1.26.4
